Define `class ModelService extends Service`
