+++ src/services/user.service.ts
@@ -1,15 +1,15 @@
 import { Op, UniqueConstraintError } from "sequelize";
 import User, { UserType, UserTypeCreation } from "../db/models/user";
 import Logged from "../log/logged.decorator";
-import Service from "./abstract.service";
+import ModelService from "./model-abstract.service";
 
 /** @private */
-interface FindQuery extends Service.FindQuery {
+interface FindQuery extends ModelService.FindQuery {
 	filter?: string;
 	filterExact?: boolean;
 }
 
-export default class UserService extends Service<User> {
+export default class UserService extends ModelService<User> {
 	@Logged({ level: "debug" })
 	protected async getRecord(id: string): Promise<User> {
 		const record = await User.findOne({
@@ -74,13 +74,13 @@ export default class UserService extends Service<User> {
 	}
 }
 
-export class UserNotFoundError extends Service.ValueNotFoundError {
+export class UserNotFoundError extends ModelService.ValueNotFoundError {
 	constructor(userID: string) {
 		super(`User "${userID}" was not found`);
 	}
 }
 
-export class UserNotUniqueError extends Service.ValueNotUniqueError {
+export class UserNotUniqueError extends ModelService.ValueNotUniqueError {
 	constructor(userLogin: string) {
 		super(`User with login "${userLogin}" already exists`);
 	}
